pyproject.toml
# Largely inspired from (pointing to precise version)
# https://github.com/thejohnfreeman/project-template-python/tree/6d04c7b3b00460bb7473246096c52cc22d403226

# black support portion inspired from (pointing to precise version)
# https://github.com/python/black/blob/4a953b7241ce5f8bcac985fa33fdf3af4f42c0de/pyproject.toml

[tool.poetry]
name = "behave4git"
version = "0.0.0"
description = "This project provides an extension (testing domain) of Behave to use Git and GitLab (and GitHub in the near future)."
license = "MIT"
homepage = "https://github.com/opinionated-digital-center/behave4git"
authors = ["Emmanuel Sciara <emmanuel.sciara@gmail.com>"]
packages = [
    { include = "behave4git" },
]
keywords = ["behave4git", "behave4cmd", "behave", "bdd"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: Jython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
]

[tool.poetry.dependencies]
python = "^3.6"
# Optional dependencies to be part of `test` extra
pytest = {version = "^5.1", optional = true}
pytest-cov = {version = "^2.8.1", optional = true}
pytest-mock = {version = "^1.13.0", optional = true}
pytest-html = {version = "^2.0.1", optional = true}
pytest-asyncio = {version = "^0.10.0", optional = true}
mock = {version = "^4.0.2", optional = true}
# Optional dependencies to be part of `bdd` extra
# Already included as non-optional dependency
# Optional dependencies to be part of `lint` extra
flake8 = {version = "^3.7", optional = true}
flake8-bugbear = {version = "^19.8.0", optional = true}
pydocstyle = {version = "^3.0", optional = true}
pylint = {version = "^2.3", optional = true}
yapf = {version = "^0.27.0", optional = true}
# Optional dependencies to be part of `type` extra
mypy = {version = "^0.740.0", optional = true}
# Optional dependencies to be part of `format` extra
isort = {version = "^4.3", optional = true}
seed-isort-config = {version = "^1.9.3", optional = true}
black = {version = "stable", optional = true, allow-prereleases = true}
# Optional dependencies to be part of `docs` extra
sphinx = {version = "^2.3.0", optional = true}
sphinx-autodoc-typehints = {version = "^1.10.3", optional = true}
sphinx-autobuild = {version = "^0.7.1", optional = true}
sphinx_rtd_theme = {version = "^0.4.3", optional = true}
# Optional dependencies to be part of `repl` extra
bpython = {version = "^0.18", optional = true}
# Non optional dependendies
#############################
# Add you dependencies here #
#############################
behave = "^1.2.6"
behave4cli = "^1.1"
PyHamcrest = "^2.0.2"
gitpython = "^3.1.1"

[tool.poetry.dev-dependencies]
tox = "^3.14"
tox-pyenv = "^1.1"
rope = "^0.14.0"

[tool.poetry.extras]
test = ["pytest", "pytest-cov", "pytest-mock", "pytest-html", "pytest-asyncio", "mock"]
lint = ["flake8", "flake8-bugbear", "pydocstyle", "pylint", "yapf"]
bdd = ["mypy"]
type = ["mypy"]
format = ["isort", "seed-isort-config", "black"]
docs= ["sphinx", "sphinx-autodoc-typehints", "sphinx-autobuild", "sphinx_rtd_theme"]
repl = ["bpython"]

[tool.isort]
# following black's doc for compatibility
# See https://github.com/psf/black#how-black-wraps-lines
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
force_grid_wrap = 0
line_length = 88
known_third_party = ["behave", "behave4cli", "git", "hamcrest", "pytest", "six", "toml"]

[tool.black]
line-length = 88
target_version = ['py37']

[build-system]
requires = ["poetry==1.0.5"]
build-backend = "poetry.masonry.api"
